Use more size_t internally, comparable with TIP #494 but then in Tk. Eliminate unnecessary type casts

+++ generic/tkTextImage.c
@@ -18,7 +18,7 @@
  */
 
 #define EI_SEG_SIZE \
-	((unsigned) (Tk_Offset(TkTextSegment, body) + sizeof(TkTextEmbImage)))
+	(Tk_Offset(TkTextSegment, body) + sizeof(TkTextEmbImage))
 
 /*
  * Prototypes for functions defined in this file:
